execution/process_data.py
"""
Example Execution Script: Process Data

This is a deterministic Python script that handles data processing.
It should be called by the orchestration layer (AI) with appropriate inputs.
"""

import os
import sys
import json
from pathlib import Path
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def validate_input(file_path):
    """Validate that the input file exists and is readable."""
    if not Path(file_path).exists():
        raise FileNotFoundError(f"Input file not found: {file_path}")
    return True

def process_data(input_file, config=None):
    """
    Process the input data file according to configuration.
    
    Args:
        input_file (str): Path to input data file
        config (dict, optional): Configuration parameters
        
    Returns:
        dict: Processing results with statistics
    """
    config = config or {}
    
    # Validate input
    validate_input(input_file)
    
    # TODO: Add actual processing logic here
    # This is just a placeholder structure
    
    results = {
        "status": "success",
        "input_file": input_file,
        "rows_processed": 0,
        "output_location": "TBD"
    }
    
    return results

def main():
    """Main entry point for the script."""
    if len(sys.argv) < 2:
        print("Usage: python process_data.py <input_file> [config_json]")
        sys.exit(1)
    
    input_file = sys.argv[1]
    config = json.loads(sys.argv[2]) if len(sys.argv) > 2 else {}
    
    try:
        results = process_data(input_file, config)
        print(json.dumps(results, indent=2))
    except Exception as e:
        error_result = {
            "status": "error",
            "error": str(e),
            "error_type": type(e).__name__
        }
        print(json.dumps(error_result, indent=2), file=sys.stderr)
        sys.exit(1)

if __name__ == "__main__":
    main()
